Create a jekyll test site for GitHub Pages
Test if GitHub pages serves up a site I never built locally.
 cd ~/Temp
 jekyll new new-blog-test-3
 cd new-blog-test-3

Modify  _config.yml for new-blog-test-3

The location for this site is: http://github.lamurakami.com/new-blog-test-3

If I had not set up a custom domain for my user repo
lamurakami.github.io when I published it the location would be:
http://lamurakami.github.io/new-blog-test-3

The default location above is a redirect created when a custom domain is
involved which in this case is because the default location for
additional GitPages for a use is as a subdirectory of the user
domain.

For a site to work as a subdirectory a base url must be spcified
in the _config.yml file for the site.

 baseurl: "" # the subpath of your site, e.g. /blog

Since I have to make modifications I will set a title.

I want a README.md with a link to the site as the minimum documentaion
for a GitHub Pages site.  This allows anyone browsing the repo to easily
get to the site.

 cp ../new-blog-test-2/README.md . # Then modify

Include README.md in about.md since I created it.
I copied the contents in since I want the README.md in the root
and not in _includes where I could use liquid {% include README.md %}

Use the Open Source Jekyll .gitignore at:
 https://github.com/github/gitignore/blob/master/Jekyll.gitignore

List files (output below)

git init
git checkout -b gh-pages
git add .
git commit

Create the new-blog-test-3 repo on GitHub without a README.md

git remote add origin git@github.com:LAMurakami/new-blog-test-3.git
git push -u origin --all
----:----|----:----|----:----|----:----|----:----|----:----|----:----|--
This is a follow up on new-blog-test-2 that I ended up building before
finding that at least the _config.yml must be modified to publish
 to new-blog-test-3 (the repo / directory name) on GitHub Pages
----:----|----:----|----:----|----:----|----:----|----:----|----:----|--
I considered adding the Site Map page but that now involves
some other changes and this test is not as simple as the first.

I wrote this commit log and now will execute the whole sequence.
----:----|----:----|----:----|----:----|----:----|----:----|----:----|--
$ find . | xargs ls -aldt --time-style=long-iso
drwxr-xr-x 8 lam staff  4096 2020-06-12 22:27 .
-rw-r--r-- 1 lam staff    52 2020-06-12 22:27 ./.gitignore
-rw-r--r-- 1 lam staff   805 2020-06-12 22:22 ./about.md
-rw-r--r-- 1 lam staff   268 2020-06-12 22:13 ./README.md
-rw-r--r-- 1 lam staff   301 2020-06-12 22:09 ./_config.yml
-rw-r--r-- 1 lam staff  1206 2020-06-12 22:03 ./_posts/2020-06-12-welcome-to-jekyll.markdown
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./_posts
-rw-r--r-- 1 lam staff   787 2020-06-12 22:03 ./_includes/icon-twitter.svg
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./_includes
-rw-r--r-- 1 lam staff   177 2020-06-12 22:03 ./_includes/icon-github.html
-rw-r--r-- 1 lam staff   180 2020-06-12 22:03 ./_includes/icon-twitter.html
-rw-r--r-- 1 lam staff   786 2020-06-12 22:03 ./_includes/head.html
-rw-r--r-- 1 lam staff   936 2020-06-12 22:03 ./_includes/footer.html
-rw-r--r-- 1 lam staff  1120 2020-06-12 22:03 ./_includes/header.html
-rw-r--r-- 1 lam staff   926 2020-06-12 22:03 ./_includes/icon-github.svg
-rw-r--r-- 1 lam staff   209 2020-06-12 22:03 ./_layouts/page.html
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./_layouts
-rw-r--r-- 1 lam staff   618 2020-06-12 22:03 ./_layouts/post.html
-rw-r--r-- 1 lam staff   248 2020-06-12 22:03 ./_layouts/default.html
-rw-r--r-- 1 lam staff   506 2020-06-12 22:03 ./index.html
-rw-r--r-- 1 lam staff 10164 2020-06-12 22:03 ./js/respond.js
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./js
-rw-r--r-- 1 lam staff 10189 2020-06-12 22:03 ./js/html5shiv.js
-rw-r--r-- 1 lam staff  1291 2020-06-12 22:03 ./feed.xml
-rw-r--r-- 1 lam staff  3355 2020-06-12 22:03 ./_sass/_syntax-highlighting.scss
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./_sass
-rw-r--r-- 1 lam staff  4299 2020-06-12 22:03 ./_sass/_layout.scss
-rw-r--r-- 1 lam staff  2909 2020-06-12 22:03 ./_sass/_base.scss
-rw-r--r-- 1 lam staff  7783 2020-06-12 22:03 ./_sass/_normalize.scss
-rw-r--r-- 1 lam staff  1063 2020-06-12 22:03 ./css/main.scss
drwxr-xr-x 2 lam staff  4096 2020-06-12 22:03 ./css
[22:28:14 Friday 06/12/2020] lam@ak19 11 files, 56K ~/Temp/new-blog-test-3